simple-webscrapper/requirements.txt
requests==2.31.0
aptos-sdk==0.8.0
python-dotenv==1.0.0